requirements.txt
python-dotenv
PyMuPDF
sqlalchemy
psycopg2-binary
